moved fields to behavior
