+++ src/redturtle/prenotazioni/behaviors/prenotazione_type_time_range/__init__.py
@@ -1,0 +1,1 @@
+# -*- coding: utf-8 -*-
