+++ src/redturtle/prenotazioni/events/configure.zcml
@@ -44,22 +44,4 @@
       handler=".prenotazione.autoconfirm"
       />
 
-  <subscriber
-      for="redturtle.prenotazioni.content.prenotazione_type.IPrenotazioneType
-           zope.lifecycleevent.IObjectCreatedEvent"
-      handler=".prenotazione_type.update_duration_from_time_range"
-      />
-
-  <subscriber
-      for="redturtle.prenotazioni.content.prenotazione_type.IPrenotazioneType
-           zope.lifecycleevent.IObjectAddedEvent"
-      handler=".prenotazione_type.update_duration_from_time_range"
-      />
-
-  <subscriber
-      for="redturtle.prenotazioni.content.prenotazione_type.IPrenotazioneType
-           zope.lifecycleevent.IObjectModifiedEvent"
-      handler=".prenotazione_type.update_duration_from_time_range"
-      />
-
 </configure>
